chore: prune deprecated harnesses
