+++ F1_24/f1_24_utils.py
@@ -1,37 +1,16 @@
 """Utility functions supporting F1 24 test script."""
 
+import os
 import re
-import sys
-from pathlib import Path
-
-PARENT_DIRECTORY = str(Path(__file__).resolve().parent.parent)
-sys.path.insert(1, PARENT_DIRECTORY)
-
-from harness_utils.paths import user_documents
-
-STEAM_GAME_ID = 2488620
-CONFIG_FILENAME = "hardware_settings_config.xml"
 
 
-def get_game_documents_dir() -> Path:
-    """Returns the F1 24 Documents directory."""
-    return user_documents(STEAM_GAME_ID) / "My Games" / "F1 24"
-
-
-def get_config_file() -> Path:
-    """Returns the F1 24 hardware settings config path."""
-    return get_game_documents_dir() / "hardwaresettings" / CONFIG_FILENAME
-
-
-def get_benchmark_results_path() -> Path:
-    """Returns the F1 24 benchmark results directory."""
-    return get_game_documents_dir() / "benchmark"
-
-
-def get_resolution(config_file: Path | None = None) -> tuple[int, int]:
+def get_resolution() -> tuple[int]:
     """Gets resolution width and height from local xml file created by game."""
+    username = os.getlogin()
+    config_path = f"C:\\Users\\{username}\\Documents\\My Games\\F1 24\\hardwaresettings"
+    config_filename = "hardware_settings_config.xml"
     resolution = re.compile(r"<resolution width=\"(\d+)\" height=\"(\d+)\"")
-    cfg = config_file or get_config_file()
+    cfg = f"{config_path}\\{config_filename}"
     height = 0
     width = 0
     with open(cfg, encoding="utf-8") as file:
