+++ forza5/forza5_utils.py
@@ -1,10 +1,9 @@
 """Utility functions for Forza Horizon 5 test script"""
 
 import re
-from pathlib import Path
 
 
-def read_resolution(config_path: str | Path) -> tuple[int, int]:
+def read_resolution(config_path: str) -> tuple[int]:
     """Gets the resolution from local file"""
     height_pattern = re.compile(r"<ResolutionHeight value=\"(\d+)\"/>")
     width_pattern = re.compile(r"<ResolutionWidth value=\"(\d+)\"/>")
@@ -16,7 +15,7 @@ def read_resolution(config_path: str | Path) -> tuple[int, int]:
             height_match = height_pattern.search(line)
             width_match = width_pattern.search(line)
             if height_match is not None:
-                height = int(height_match.group(1))
+                height = height_match.group(1)
             if width_match is not None:
-                width = int(width_match.group(1))
+                width = width_match.group(1)
     return (width, height)
