--- zdeprecated/F1_22/F1_22.toml
@@ -1,5 +1,0 @@
-friendly_name = "F1_22"
-py_script = "f1.py"
-executable = "F1_22.exe"
-presentmon_enabled = true
-presentmon_ignore_exit = true
