--- zdeprecated/F1_22/manifest.yaml
@@ -1,10 +1,0 @@
-friendly_name: "F1_22"
-executable: "f1.py"
-process_name: "F1_22.exe"
-hidden: 0
-output_dir: "run"
-options:
-  - name: kerasHost
-    type: input
-  - name: kerasPort
-    type: input
