--- zdeprecated/F1_23/F1_23.toml
@@ -1,5 +1,0 @@
-friendly_name = "F1_23"
-py_script = "f1_23.py"
-executable = "F1_23.exe"
-presentmon_enabled = true
-presentmon_ignore_exit = true
