--- zdeprecated/F1_23/manifest.yaml
@@ -1,10 +1,0 @@
-friendly_name: "F1_23"
-executable: "f1_23.py"
-process_name: "F1_23.exe"
-hidden: 0
-output_dir: "run"
-options:
-  - name: kerasHost
-    type: input
-  - name: kerasPort
-    type: input
