--- zdeprecated/README.md
@@ -1,6 +1,0 @@
-# Deprecated Harnesses
-> **NOTE:** Tests in this directory are deprecated; They are  **not** guaranteed to work!
-
-These harnesses are, as the name of directory suggests, deprecated. This means that they are not actively maintained and are not guaranteed to work. They may have bugs, unfinished code, and/or utilize approaches that we have moved away from as we continue to evolve MarkBench and explore newer and better test harnesses. 
-
-The deprecated harnesses we have placed here currently are meant for games or software that we are not actively testing, but have decided to keep around for one reason or another for the time being. They may be revisited at some point or deleted outright should we decide they no longer have value. For example, we may consider revisiting a harness in the future and would like to keep the original script as a reference for when we go back to give it an update.
